complete working example
Here's a working example of the project we started during the lecture.
After we were finished I added a call to the FHIR API to search for
patients. I added the access_token we got from our  OpenID Connect
provider to the Authorization HTTP header, in addition to the Open DIPS
subscription key needed to call the FHIR API.